Split ScalaDslTest in multiple files to ease maintenance
